refactor: thin camera-ready campaign legacy facade (#3385)
